Updated: ffinput.c changed log level of an unimportant error from LOG_ERROR to LOG_WARNING. Added: Equalizer settings are now saved accross application instances.

+++ opusapp/src/main/java/net/opusapp/player/core/service/providers/AbstractMediaManager.java
@@ -264,7 +264,7 @@ public interface AbstractMediaManager {
 
         public boolean equalizerIsEnabled();
         public long equalizerSetEnabled(boolean enabled);
-        public long equalizerBandSetGain(int band, int gain);
+        public long equalizerBandSetGain(int band, long gain);
         public long equalizerBandGetGain(int band);
         public boolean equalizerApplyProperties();
